Add io.Pipe compatible interface
The ring buffer can be used as a compatible, but *asynchronous* replacement of `io.Pipe`.

That means that Reads and Writes will go to the ring buffer.
Writes will complete as long as the data fits within the ring buffer.

Reads will attempt to satisfy reads with data from the ring buffer.
Only if the ring buffer is empty will the read block.

In the common case, where the Read and Write side can run concurrently,
it is safe to replace `io.Pipe()` with `(*Ringbuffer).Pipe()`.