Add optional timeout to blocking reads/writes
This can be helpful to time out on blocked operations where the opposite end completely stops accepting or delivering data.

This is not a guarantee that the Read/Write will return within that deadline, since there may be several operations needed to fulfil a single operation.

Also fixes up some minor grammar stuff.