Add separate read/write deadlines
Allows to control deadlines on each side of the buffer.

An example would be a webserver that receives data from slow remotes.
The data may be slow or intermittent, but we expect the buffer to be cleared within reasonable time.

If the ringbuffer is reading from the request, we can therefore set a long (or no) timeout on reads,
but set a short a 5 second on writes, since we expect our backend to handle data within this timeframe.